requirements.txt
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
decorator==5.1.1
exceptiongroup==1.1.1
executing==1.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
ipython==8.13.1
jedi==0.18.2
kaitaistruct==0.10
matplotlib-inline==0.1.6
outcome==1.2.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==1.0.0
python-vlc==3.0.18122
PyVirtualDisplay==3.0
requests==2.29.0
selenium==4.9.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
wsproto==1.2.0
zstandard==0.21.0
